+++ Test/Program.cs
@@ -17,12 +17,6 @@ namespace Test
                 HetznerCloudClient hetznerCloudClient = new HetznerCloudClient("ApiKey");
 
                 hetznerCloudClient = new HetznerCloudClient(await File.ReadAllTextAsync("D:\\HetznerApiKey.txt"));
-
-                List<Action> list = await hetznerCloudClient.VolumeAction.GetAllActions();
-
-                long actionId = 1236866267;
-
-                Action action = await hetznerCloudClient.VolumeAction.GetAction(actionId);
             }
             catch (Exception ex)
             {